task 3.1. fixed
